requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
groq>=0.9.0
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
